+++ chouette-core/src/main/java/fr/certu/chouette/model/neptune/StopArea.java
@@ -202,8 +202,8 @@ public class StopArea extends NeptuneLocalizedObject
    @Setter
    @ManyToMany
    @JoinTable(name = "stop_areas_stop_areas",
-         joinColumns = { @JoinColumn(name = "child_id", nullable = false, updatable = false) },
-         inverseJoinColumns = { @JoinColumn(name = "parent_id", nullable = false, updatable = false) })
+         joinColumns = { @JoinColumn(name = "parent_id", nullable = false, updatable = false) },
+         inverseJoinColumns = { @JoinColumn(name = "child_id", nullable = false, updatable = false) })
    private List<StopArea> routingConstraintAreas = new ArrayList<StopArea>(0);
 
    @Getter
@@ -949,6 +949,15 @@ public class StopArea extends NeptuneLocalizedObject
          {
             addRoutingConstraintLineId(line.getObjectId());
          }
+      }
+      
+      if (getRoutingConstraintAreas() != null)
+      {
+          for (StopArea child : getRoutingConstraintAreas())
+          {
+             addContainedStopId(child.getObjectId());
+          }
+    	  
       }
 
       // connectionlinks are not mapped by hibernate ; must fill with 2