+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * L'Import GTFS accepte des routes sans colonne agencyId (Mantis 22665)
   * L'Export GTFS sort en erreur sur objet à exportee incomplêt (Mantis 24484)
   * L'Export CSV sort un message inexploitable sur informations incomplètes (Mantis 24485)
+  * L'import NeTEx n'importe pas les ITL (Mantis 20889)
   
 # Version 2.3.0 (06/03/14, en cours de qualification)
 * migration technique